chore: class renames
