+++ tee-aps/sealed-bid-auction/src/api/http/SealedBidAuctionController.ts
@@ -1,6 +1,6 @@
 import type {Request, Response} from "express";
 
-export class AuctionController {
+export class SealedBidAuctionController {
 
     async open(req: Request, res: Response): Promise<void> {
         res.status(501);
